Allow deleting of publishables
It is possible that the original author was not planning on allowing
publications to ever be deleted, in favor of "unpublishing" them, or
archiving them, instead.  But all that jazz with
publication/unpublication/archive dates and times doesn't seem to be
implemented at all, and just allowing things to be deleted seems
appropriate in many cases and was certainly easier to implement.

But you can still look forward to unpublish/archive options at some
point.